refactor: car와 race 패키지 분리
